requirements.txt
fastapi==0.95.1
fastapi-utils==0.2.1
uvicorn==0.22.0
torch
transformers
pyyaml
requests